requirements.txt
mysql-connector-python==9.2.0
streamlit==1.41.1